requirements.txt
pymupdf>=1.24,<2
pdfplumber>=0.11,<1
psycopg[binary]>=3.2,<4
python-dotenv>=1.0,<2

